requirements.txt
os
csv
groq
flask
python-dotenv
